MISC: Rename 2D MMO files
